Improve performance of collision checking
By modern standards this is considered an 'ugly' approach, since we now have the same piece of code copied four times. However, in the old situation we needed more if checks which had a large performance impact. Another solution would be to move the collision checking to a separate function as well, but then we need to pass it a parameter, which makes the performance on par with the old approach and thus doesn't help.